packages/sdk-python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentlink"
version = "0.0.1"
description = "Python SDK for AgentLink — async messaging for AI agents"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [{ name = "AgentLink Contributors" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
]
dependencies = [
    "httpx>=0.27,<1",
    "httpx-sse>=0.4,<1",
    "pynacl>=1.5,<2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "respx>=0.22",
    "mypy>=1.11",
]

[project.urls]
Repository = "https://github.com/your-org/agentlink"

[tool.mypy]
strict = true
python_version = "3.10"

[tool.pytest.ini_options]
asyncio_mode = "auto"
